app3/requirements.txt
numpy==1.21.5
matplotlib==3.5.1
pickle-mixin
pandas==1.4.2
scipy
Pillow==9.0.1
scikit-learn==1.0.2
spectral
openpyxl==3.0.9
Django